Updating superuser creation
This ran through without generating an error so while I can't be sure its
working exactly as expected I think its a good enough update to commit. I will
need a clean instance to know if it is working as expected. What we do know is
that this code doesn't update the existing superuser.