Update systemd unit
This brings in some commented suggestions from upstream doco - which will
probably be good to add - and expands the loging capabilities of the unit.

Add a call to reload systemd . This should be fixed to only run on unit change
and use the built in ansible support.

Whitespace fix in systemd config at the same time