Update configuration for T11
This is the minimum changes required to make the role forward port to Tendenci
11. It would be nice to make quite a few enhancements to configuration handling
at a future date.

Some changes (updated secret key lengths, default port) aren't strictly related
but are being updated as a result of this upgrade and it seems reasonable to
include them here.

Also create logs directory - logs are now baked in to Tendenci so need to be
configured up front.