Remove pre defined secrets
First pass at fixing #11 using two different methods.
For Django / Tendenci values (tendenci_site_secret_key and
tendenci_site_settings_key) we attempt to generate the key on the fly if its
missing.
For the postgresql user we simply fail, as this value has to be coordinated
with the chosen postgresql role and can't be made up on the spot.

Neither change has been proeprly tested yet, so view this as experimental.