Convert repository to housing Ansible collection
I'm about to substantially rework this role , changing dependencies and related
functionality. Because Collections have come in as a preferred method of
distribution since I last updated this role I am taking this opportunity to
wrap the role up in a collection.

Apologies to anyone using this direct from the repository who is now broken, I
hope the inconvenience isn't too great.